plugindatabase: Split into header and header-only-implementation
